refactor(desktop): split the install-report tests by concern
The redaction cases grew the single test file past the desktop 1000-line
ratchet, which is a hard cap for a file that does not exist on main. They
are a self-contained concern, so they move to their own module alongside
the shared harness, following the split migration.rs already uses.

No test changes: 37 before, 37 after.
